feat(k8s): add postgres deployment and apply script
Add a new file 'deploy-postgres.yml' to the k8s directory, which contains the
definition for deploying a postgres container to the k8s cluster. Also, add a
new file 'apply.sh' to the k8s directory, which contains a script to apply the
postgres container to the k8s cluster and display information about the
deployments, pods, and services.

The 'deploy-postgres.yml' file includes the deployment configuration for a
postgres container with the following environment variables:
- POSTGRES_USER: postgres
- POSTGRES_PASSWORD: postgres
- POSTGRES_DB: address-book-db
The container listens on port 5432.

The script 'apply.sh' uses 'kubectl apply' command to deploy the postgres
container to the k8s cluster and then displays information about the
deployments, pods, and services using 'kubectl get' commands.